requirements.txt
asgiref==3.8.1
black==25.1.0
boto3==1.34.75
click==8.1.8
colorama==0.4.6
dj-database-url==2.2.0
Django==5.1.6
django-cors-headers==4.7.0
django-storages==1.14.2
djangorestframework==3.15.2
gunicorn==21.2.0
iniconfig==2.0.0
mypy-extensions==1.0.0
nulltype==2.3.1
packaging==24.2
pathspec==0.12.1
plaid-python==29.0.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
stripe==7.12.0
supabase>=2.3.1
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.5.0
